backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
django-storages==1.14.2
boto3==1.34.0
drf-yasg==1.21.7
django-filter==23.5